requirements.txt
bleach>=6.0.0
fastapi>=0.100.0
gssapi>=1.9.0
httpx>=0.27.0
Jinja2>=3.1.0
ldap3>=2.9.1
markdown>=3.4.0
prometheus-client>=0.17.0
psutil>=5.9.0
pynvml>=11.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pypdf>=5.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
redis>=4.0.0
requests>=2.31.0
uvicorn[standard]>=0.23.0
